+++ src/components/event/Body.tsx
@@ -1,4 +1,4 @@
-import Link from "next/link";
+import { WrapLinks } from "../WrapLinks";
 
 type Props = { description: string };
 
@@ -14,49 +14,4 @@ export const Body = ({ description }: Props) => {
       ))}
     </div>
   );
-};
-
-type WrapLinksProps = { text: string };
-
-const WrapLinks = ({ text }: WrapLinksProps) => {
-  const parts = text.split(/\s+/g);
-
-  return (
-    <span>
-      {parts.map((part, index) => {
-        if (part.startsWith("http")) {
-          const url = safeParseUrl(part);
-          if (!url) {
-            return <span key={`wrap-link-${index}`}> {part} </span>;
-          }
-
-          return (
-            <Link
-              className="hover:underline"
-              key={`wrap-link-${index}`}
-              rel="noopener noreferrer"
-              target="_blank"
-              href={url.href}
-            >
-              {part}
-            </Link>
-          );
-        }
-
-        return <span key={`wrap-link-${index}`}> {part} </span>;
-      })}
-    </span>
-  );
-};
-
-const safeParseUrl = (url: string) => {
-  if (!url.startsWith("http")) {
-    return null;
-  }
-
-  try {
-    return new URL(encodeURI(url));
-  } catch (error) {
-    return null;
-  }
 };
